+++ Assets/VRM/Editor/VrmTopMenu.cs
@@ -55,9 +55,6 @@ namespace VRM
 
         [MenuItem(DevelopmentMenuPrefix + "/Create UnityPackage", priority = 34)]
         private static void CreateUnityPackage() => VRMExportUnityPackage.CreateUnityPackageWithoutBuild();
-
-        [MenuItem(DevelopmentMenuPrefix + "/Export ShaderProps Code", priority = 35)]
-        private static void ExportShaderPropertyCode() => ShaderPropMenu.PreExport();
 #endif
     }
 }
