--- Assets/VRMShaders/VRM/IO/Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 36a5a07493511bd49b1ad9dbad2683d2
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
