--- Assets/VRMShaders/VRM/IO/Editor/ShaderPropMenu.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5eb12c186c6337e4db278b5f01d47cae
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
