--- Assets/VRMShaders/VRM/IO/Editor/VRMShaders.VRM.IO.Editor.asmdef
@@ -1,17 +1,0 @@
-{
-    "name": "VRMShaders.VRM.IO.Editor",
-    "references": [
-        "GUID:301b251fd9834274c9228e0532f444f7"
-    ],
-    "includePlatforms": [
-        "Editor"
-    ],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": false,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
