--- Assets/VRMShaders/VRM/IO/Editor/VRMShaders.VRM.IO.Editor.asmdef.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: bc66ece0f33b52446a0830c05781d4db
-AssemblyDefinitionImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
