--- Assets/VRMShaders/VRM/IO/Runtime/GLTF.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d9c97ad7f5bbcac489a47a2f34dfff00
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
