--- Assets/VRMShaders/VRM/IO/Runtime/GLTF/PreExportShaders_GLTF.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 90dcef1978c51e74386b76d77689dc82
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
