--- Assets/VRMShaders/VRM/IO/Runtime/GLTF/Standard.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 54043e349b047bf4b8f127cd919a757d
-timeCreated: 1533542890
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
