--- Assets/VRMShaders/VRM/IO/Runtime/GLTF/UniGLTF_UniUnlit.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: ad4b6b115b83ecd48a513f697afc95f0
-timeCreated: 1537860074
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
