--- Assets/VRMShaders/VRM/IO/Runtime/GLTF/Unlit_Color.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 4f91421c5846d5d48933d2ee4ffeeceb
-timeCreated: 1535186213
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
