--- Assets/VRMShaders/VRM/IO/Runtime/GLTF/Unlit_Texture.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 8273e1e61ad8e914baae94d06836f2ad
-timeCreated: 1535186213
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
