--- Assets/VRMShaders/VRM/IO/Runtime/GLTF/Unlit_Transparent.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: b98a8ee8ca13abb43809305cc4e5571a
-timeCreated: 1535186213
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
