--- Assets/VRMShaders/VRM/IO/Runtime/GLTF/Unlit_Transparent_Cutout.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 844490f13976543478d82efe28251941
-timeCreated: 1535186301
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
