+++ Assets/VRMShaders/VRM/IO/Runtime/PreShaderPropExporter.cs
@@ -1,61 +1,63 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection;
-using UnityEngine;
-#if UNITY_EDITOR
-using UnityEditor;
-#endif
-
-
-namespace UniGLTF.ShaderPropExporter
+﻿namespace UniGLTF.ShaderPropExporter
 {
-    public class PreExportShadersAttribute : Attribute { }
-    public class PreExportShaderAttribute : Attribute { }
-
-    public struct SupportedShader
-    {
-        public string TargetFolder;
-        public string ShaderName;
-
-        public SupportedShader(string targetFolder, string shaderName)
-        {
-            TargetFolder = targetFolder;
-            ShaderName = shaderName;
-        }
-    }
-
     /// <summary>
-    /// 今はMToon専用
+    /// TODO: MToon にひきとってもらう
     /// </summary>
-    public static partial class PreShaderPropExporter
+    public static class PreShaderPropExporter
     {
-        static Dictionary<string, ShaderProps> m_shaderPropMap;
-
         const string ShaderName = "VRM/MToon";
-
         public static ShaderProps GetPropsForMToon()
         {
-            if (m_shaderPropMap == null)
-            {
-                m_shaderPropMap = new Dictionary<string, ShaderProps>();
-                foreach (var prop in typeof(PreShaderPropExporter).GetProperties(BindingFlags.Static | BindingFlags.Public | BindingFlags.NonPublic))
-                {
-                    if (prop.GetCustomAttributes(typeof(PreExportShaderAttribute), true).Any())
-                    {
-                        var kv = (KeyValuePair<string, ShaderProps>)prop.GetValue(null, null);
-                        m_shaderPropMap.Add(kv.Key, kv.Value);
-                    }
-                }
-            }
-
-            ShaderProps props;
-            if (m_shaderPropMap.TryGetValue(ShaderName, out props))
-            {
-                return props;
-            }
-
-            throw new Exception("arienai");
+            return VRM_MToon;
         }
+
+        static ShaderProps VRM_MToon =
+                    new ShaderProps
+                    {
+                        Properties = new ShaderProperty[]{
+new ShaderProperty("_Cutoff", ShaderPropertyType.Range)
+,new ShaderProperty("_Color", ShaderPropertyType.Color)
+,new ShaderProperty("_ShadeColor", ShaderPropertyType.Color)
+,new ShaderProperty("_MainTex", ShaderPropertyType.TexEnv)
+,new ShaderProperty("_ShadeTexture", ShaderPropertyType.TexEnv)
+,new ShaderProperty("_BumpScale", ShaderPropertyType.Float)
+,new ShaderProperty("_BumpMap", ShaderPropertyType.TexEnv)
+,new ShaderProperty("_ReceiveShadowRate", ShaderPropertyType.Range)
+,new ShaderProperty("_ReceiveShadowTexture", ShaderPropertyType.TexEnv)
+,new ShaderProperty("_ShadingGradeRate", ShaderPropertyType.Range)
+,new ShaderProperty("_ShadingGradeTexture", ShaderPropertyType.TexEnv)
+,new ShaderProperty("_ShadeShift", ShaderPropertyType.Range)
+,new ShaderProperty("_ShadeToony", ShaderPropertyType.Range)
+,new ShaderProperty("_LightColorAttenuation", ShaderPropertyType.Range)
+,new ShaderProperty("_IndirectLightIntensity", ShaderPropertyType.Range)
+,new ShaderProperty("_RimColor", ShaderPropertyType.Color)
+,new ShaderProperty("_RimTexture", ShaderPropertyType.TexEnv)
+,new ShaderProperty("_RimLightingMix", ShaderPropertyType.Range)
+,new ShaderProperty("_RimFresnelPower", ShaderPropertyType.Range)
+,new ShaderProperty("_RimLift", ShaderPropertyType.Range)
+,new ShaderProperty("_SphereAdd", ShaderPropertyType.TexEnv)
+,new ShaderProperty("_EmissionColor", ShaderPropertyType.Color)
+,new ShaderProperty("_EmissionMap", ShaderPropertyType.TexEnv)
+,new ShaderProperty("_OutlineWidthTexture", ShaderPropertyType.TexEnv)
+,new ShaderProperty("_OutlineWidth", ShaderPropertyType.Range)
+,new ShaderProperty("_OutlineScaledMaxDistance", ShaderPropertyType.Range)
+,new ShaderProperty("_OutlineColor", ShaderPropertyType.Color)
+,new ShaderProperty("_OutlineLightingMix", ShaderPropertyType.Range)
+,new ShaderProperty("_UvAnimMaskTexture", ShaderPropertyType.TexEnv)
+,new ShaderProperty("_UvAnimScrollX", ShaderPropertyType.Float)
+,new ShaderProperty("_UvAnimScrollY", ShaderPropertyType.Float)
+,new ShaderProperty("_UvAnimRotation", ShaderPropertyType.Float)
+,new ShaderProperty("_MToonVersion", ShaderPropertyType.Float)
+,new ShaderProperty("_DebugMode", ShaderPropertyType.Float)
+,new ShaderProperty("_BlendMode", ShaderPropertyType.Float)
+,new ShaderProperty("_OutlineWidthMode", ShaderPropertyType.Float)
+,new ShaderProperty("_OutlineColorMode", ShaderPropertyType.Float)
+,new ShaderProperty("_CullMode", ShaderPropertyType.Float)
+,new ShaderProperty("_OutlineCullMode", ShaderPropertyType.Float)
+,new ShaderProperty("_SrcBlend", ShaderPropertyType.Float)
+,new ShaderProperty("_DstBlend", ShaderPropertyType.Float)
+,new ShaderProperty("_ZWrite", ShaderPropertyType.Float)
+                        }
+                    };
     }
 }
