--- Assets/VRMShaders/VRM/IO/Runtime/VRM/PreExportShaders_VRM.cs
@@ -1,16 +1,0 @@
-﻿namespace UniGLTF.ShaderPropExporter
-{
-    public static partial class PreExportShaders
-    {
-        const string VRM_TARGET_FOLDER = "VRM";
-        [PreExportShaders]
-        public static SupportedShader[] VRMSupportedShaders = new SupportedShader[]
-        {
-            new SupportedShader(VRM_TARGET_FOLDER, "VRM/MToon"),
-            new SupportedShader(VRM_TARGET_FOLDER, "VRM/UnlitTexture"),
-            new SupportedShader(VRM_TARGET_FOLDER, "VRM/UnlitCutout"),
-            new SupportedShader(VRM_TARGET_FOLDER, "VRM/UnlitTransparent"),
-            new SupportedShader(VRM_TARGET_FOLDER, "VRM/UnlitTransparentZWrite"),
-        };
-    }
-}
