--- Assets/VRMShaders/VRM/IO/Runtime/VRM/PreExportShaders_VRM.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 625b5ee8b5811dc4a915a2fbb2cb319d
-timeCreated: 1533035131
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
