--- Assets/VRMShaders/VRM/IO/Runtime/VRM/VRM_MToon.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 4629d794c8969c141a4724e182af082e
-timeCreated: 1533542890
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
