--- Assets/VRMShaders/VRM/IO/Runtime/VRM/VRM_UnlitCutout.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 611b546ea471ad34cb7d94740c63b558
-timeCreated: 1533542890
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
