--- Assets/VRMShaders/VRM/IO/Runtime/VRM/VRM_UnlitTexture.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 22a8083880389b3498f421e6a5c340d5
-timeCreated: 1533542890
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
