--- Assets/VRMShaders/VRM/IO/Runtime/VRM/VRM_UnlitTransparent.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 140d6538826e0eb448929d3e4bb2f1cd
-timeCreated: 1533542890
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
