--- Assets/VRMShaders/VRM/IO/Runtime/VRM/VRM_UnlitTransparentZWrite.cs
@@ -1,26 +1,0 @@
-using System.Collections.Generic;
-
-
-namespace UniGLTF.ShaderPropExporter
-{
-    public static partial class PreShaderPropExporter
-    {
-        [PreExportShader]
-        static KeyValuePair<string, ShaderProps> VRM_UnlitTransparentZWrite 
-        {
-            get 
-            {
-                return new KeyValuePair<string, ShaderProps>(
-                    "VRM/UnlitTransparentZWrite",
-                    new ShaderProps
-                    {
-                        Properties = new ShaderProperty[]{
-new ShaderProperty("_MainTex", ShaderPropertyType.TexEnv)
-
-                        }
-                    }
-                );
-            }
-        }
-    }
-}
