--- Assets/VRMShaders/VRM/IO/Runtime/VRM/VRM_UnlitTransparentZWrite.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 165ec79b7aac1564a850fb3d3d19396e
-timeCreated: 1533542890
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
